+++ src/components/restaurant/ReviewForm.jsx
@@ -1,63 +1,11 @@
-import React, { useReducer } from 'react';
+import React from 'react';
 import Counter from './Counter';
-
-const initialState = {
-  name: '',
-  rating: 0,
-  comment: '',
-};
-
-const ACTION_TYPES = {
-  SET_NAME: 'SET_NAME',
-  SET_RATING: 'SET_RATING',
-  SET_COMMENT: 'SET_COMMENT',
-  RESET: 'RESET',
-};
-
-const reviewReducer = (state, action) => {
-  switch (action.type) {
-    case ACTION_TYPES.SET_NAME:
-      return { ...state, name: action.payload };
-    case ACTION_TYPES.SET_RATING:
-      return { ...state, rating: action.payload };
-    case ACTION_TYPES.SET_COMMENT:
-      return { ...state, comment: action.payload };
-    case ACTION_TYPES.RESET:
-      return initialState;
-    default:
-      return state;
-  }
-};
+import { useReviewForm } from '../../hooks/useReviewForm';
 
 const ReviewForm = ({ addReview }) => {
-  const [state, dispatch] = useReducer(reviewReducer, initialState);
+  const { state, handleChange, handleTextChange, handleRatingChange, handleClear } =
+    useReviewForm();
 
-  const handleChange = (e) => {
-    dispatch({
-      type: ACTION_TYPES.SET_NAME,
-      payload: e.target.value,
-    });
-  };
-
-  const handleTextChange = (e) => {
-    dispatch({
-      type: ACTION_TYPES.SET_COMMENT,
-      payload: e.target.value,
-    });
-  };
-
-  const handleRatingChange = (newValue) => {
-    dispatch({
-      type: ACTION_TYPES.SET_RATING,
-      payload: Number(newValue),
-    });
-  };
-
-  const handleClear = () => {
-    dispatch({
-      type: ACTION_TYPES.RESET,
-    });
-  };
   const handleSubmit = (e) => {
     e.preventDefault();
 
@@ -68,9 +16,7 @@ const ReviewForm = ({ addReview }) => {
 
     if (addReview) {
       addReview(state);
-      dispatch({
-        type: ACTION_TYPES.RESET,
-      });
+      handleClear();
     }
   };
